add parallel downloading for sun3d downloader
